backend/requirements.txt
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
fastapi==0.121.3
uvicorn==0.38.0
pydantic==2.12.4
python-decouple==3.8
redis==5.3.1
supabase==1.0.1
httpx==0.23.3
httpcore==0.16.3
h11==0.14.0
anyio==4.11.0
langdetect==1.0.9
requests==2.32.5